+++ src/main/java/tech/reliab/course/mikhaylyukovada/bank/service/impl/BankServiceImpl.java
@@ -1,8 +1,14 @@
 package tech.reliab.course.mikhaylyukovada.bank.service.impl;
 
-import tech.reliab.course.mikhaylyukovada.bank.entity.Bank;
+import tech.reliab.course.mikhaylyukovada.bank.entity.*;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.FailedLoanException;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.ErrorSumException;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.IdNotFoundException;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.WrongNameException;
 import tech.reliab.course.mikhaylyukovada.bank.repository.*;
-import tech.reliab.course.mikhaylyukovada.bank.service.BankService;
+import tech.reliab.course.mikhaylyukovada.bank.service.*;
+import tech.reliab.course.mikhaylyukovada.bank.utils.BankComparator;
+
 import java.util.List;
 
 /**
@@ -12,15 +18,26 @@ public class BankServiceImpl implements BankService {
     private static BankServiceImpl INSTANCE;
 
     public static BankServiceImpl getInstance() {
-        if (INSTANCE == null) { INSTANCE = new BankServiceImpl(); }
+        if (INSTANCE == null) {
+            INSTANCE = new BankServiceImpl();
+        }
 
         return INSTANCE;
     }
 
     private final BankRepository bankRepository = BankRepository.getInstance();
 
+    private static final double MAX_INTEREST_RATE = 20.d;
+    private static final int MAX_MONEY = 1_000_000;
+    private static final int MIN_CLIENT_RATING = 5000;
+    private static final int TOP_BANK_RATING = 50;
+
     @Override
     public Bank addObject(Bank bank) {
+        if (bank.getName().isBlank()) {
+            throw new WrongNameException();
+        }
+
         return bankRepository.add(bank);
     }
 
@@ -31,11 +48,19 @@ public class BankServiceImpl implements BankService {
 
     @Override
     public boolean deleteObjectById(Long id) {
+        if (id == null) {
+            throw new IdNotFoundException();
+        }
+
         return bankRepository.deleteById(id);
     }
 
     @Override
     public Bank getObjectById(Long id) {
+        if (id == null) {
+            throw new IdNotFoundException();
+        }
+
         return bankRepository.findById(id);
     }
 
@@ -107,5 +132,123 @@ public class BankServiceImpl implements BankService {
             }
         }
 
+    }
+
+    @Override
+    public Long getLoan(Long userId, Double creditSum) throws FailedLoanException {
+        AtmService atmService = AtmServiceImpl.getInstance();
+        PaymentAccountService paymentAccountService = PaymentAccountServiceImpl.getInstance();
+        EmployeeService employeeService = EmployeeServiceImpl.getInstance();
+        BankOfficeService bankOfficeService = BankOfficeServiceImpl.getInstance();
+        UserService userService = UserServiceImpl.getInstance();
+        CreditAccountService creditAccountService = CreditAccountServiceImpl.getInstance();
+
+        if (creditSum <= 0) {
+            throw new ErrorSumException();
+        }
+
+        User user = userService.getObjectById(userId);
+        var banks = bankRepository.findAll().stream().sorted(new BankComparator()).toList();
+
+        for (int i = banks.size() - 1; i >= 0; i--) {
+            if (user.getCreditRating() < MIN_CLIENT_RATING && banks.get(i).getBankRating() > TOP_BANK_RATING) {
+                continue;
+            }
+
+            var offices = bankOfficeService.getAllBankOfficesByBankId(banks.get(i).getId());
+
+            for (BankOffice office : offices) {
+                if (!office.getGettingLoan() || office.getTotalMoney().compareTo(creditSum) < 0) {
+                    continue;
+                }
+
+                var employee = employeeService.getEmployeeWithLoan(office.getId());
+                if (employee == null) {
+                    continue;
+                }
+
+                var atms = atmService.getAllBankAtmsByOffice(office.getId());
+                for (BankAtm atm : atms) {
+                    if (!atm.getAcceptingMoney() || atm.getTotalMoney().compareTo(creditSum) < 0) {
+                        continue;
+                    }
+
+                    PaymentAccount paymentAccount = paymentAccountService.getPaymentAccount(banks.get(i), user);
+                    if (paymentAccount == null) paymentAccountService.getNewPaymentAccount(banks.get(i), user);
+                    int monthNumber = (int) (creditSum / user.getMonthlyIncome());
+                    CreditAccount creditAccount = creditAccountService.createCreditAccount(banks.get(i), user,
+                            paymentAccount, employee, creditSum, monthNumber);
+
+                    atmService.withdrawMoney(atm.getId(), creditSum);
+
+                    return creditAccountService.addObject(creditAccount).getId();
+                }
+
+            }
+        }
+        throw new FailedLoanException();
+    }
+
+    @Override
+    public Long getLoanInCurrentBank(Long userId, Long bankId, Double creditSum) throws FailedLoanException {
+        AtmService atmService = AtmServiceImpl.getInstance();
+        PaymentAccountService paymentAccountService = PaymentAccountServiceImpl.getInstance();
+        EmployeeService employeeService = EmployeeServiceImpl.getInstance();
+        BankOfficeService bankOfficeService = BankOfficeServiceImpl.getInstance();
+        UserService userService = UserServiceImpl.getInstance();
+        CreditAccountService creditAccountService = CreditAccountServiceImpl.getInstance();
+
+        if (creditSum <= 0) {
+            throw new ErrorSumException();
+        }
+
+        User user = userService.getObjectById(userId);
+        Bank bank = bankRepository.findById(bankId);
+
+        if (user.getCreditRating() < MIN_CLIENT_RATING && bank.getBankRating() > TOP_BANK_RATING) {
+            throw new FailedLoanException("Credit rating too low ");
+        }
+
+        var offices = bankOfficeService.getAllBankOfficesByBankId(bank.getId());
+
+        var isEnoughMoney = false;
+
+        for (BankOffice office : offices) {
+            if (!office.getGettingLoan() || office.getTotalMoney().compareTo(creditSum) < 0) {
+                continue;
+            }
+
+            var employee = employeeService.getEmployeeWithLoan(office.getId());
+            if (employee == null) {
+                continue;
+            }
+
+            var atms = atmService.getAllBankAtmsByOffice(office.getId());
+            for (BankAtm atm : atms) {
+                if (!atm.getAcceptingMoney() || atm.getTotalMoney().compareTo(creditSum) < 0) {
+                    isEnoughMoney = true;
+                    continue;
+                }
+
+                PaymentAccount paymentAccount = paymentAccountService.getPaymentAccount(bank, user);
+
+                if (paymentAccount == null) {
+                    throw new FailedLoanException("User doesn't have credit account");
+                }
+
+                int monthNumber = (int) (creditSum / user.getMonthlyIncome());
+                CreditAccount creditAccount = creditAccountService.createCreditAccount(bank, user,
+                        paymentAccount, employee, creditSum, monthNumber);
+
+                atmService.withdrawMoney(atm.getId(), creditSum);
+
+                return creditAccountService.addObject(creditAccount).getId();
+            }
+
+        }
+
+        if (!isEnoughMoney) throw new FailedLoanException("Not enough money in office");
+
+        throw new FailedLoanException("Just Error");
     }
 }
