+++ src/main/java/tech/reliab/course/mikhaylyukovada/bank/service/impl/CreditAccountServiceImpl.java
@@ -1,8 +1,13 @@
 package tech.reliab.course.mikhaylyukovada.bank.service.impl;
 
-import tech.reliab.course.mikhaylyukovada.bank.entity.CreditAccount;
+import tech.reliab.course.mikhaylyukovada.bank.entity.*;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.IdNotFoundException;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.WrongNameException;
+import tech.reliab.course.mikhaylyukovada.bank.exceptions.WrongPaymentException;
 import tech.reliab.course.mikhaylyukovada.bank.repository.CreditAccountRepository;
 import tech.reliab.course.mikhaylyukovada.bank.service.CreditAccountService;
+
+import java.time.LocalDate;
 import java.util.List;
 
 /**
@@ -17,6 +22,11 @@ public class CreditAccountServiceImpl implements CreditAccountService {
         return INSTANCE;
     }
 
+    public CreditAccount createCreditAccount(Bank bank, User user, PaymentAccount paymentAccount, Employee employee, Double sum, int monthNumber) {
+        return new CreditAccount(user, bank.getName(), LocalDate.now(), monthNumber, sum,
+                bank.getInterestRate(), employee, paymentAccount);
+    }
+
     private final CreditAccountRepository creditAccountRepository = CreditAccountRepository.getInstance();
 
     @Override
@@ -24,6 +34,11 @@ public class CreditAccountServiceImpl implements CreditAccountService {
         creditAccount.setEndDate(creditAccount.getStartDate().plusMonths(creditAccount.getMonthsNumber()));
 
         var monthlyPayment = creditAccount.getCreditAmount() / creditAccount.getMonthsNumber().doubleValue() * creditAccount.getInterestRate();
+
+        if (monthlyPayment <= 0) {
+            throw new WrongPaymentException();
+        }
+
         creditAccount.setMonthlyPayment(monthlyPayment);
 
         return creditAccountRepository.add(creditAccount);
@@ -36,11 +51,19 @@ public class CreditAccountServiceImpl implements CreditAccountService {
 
     @Override
     public boolean deleteObjectById(Long id) {
+        if (id == null) {
+            throw new IdNotFoundException();
+        }
+
         return creditAccountRepository.deleteById(id);
     }
 
     @Override
     public CreditAccount getObjectById(Long id) {
+        if (id == null) {
+            throw new IdNotFoundException();
+        }
+
         return creditAccountRepository.findById(id);
     }
